requirements/common.txt
djangorestframework==3.9.1
Django==1.11.8
mysqlclient==1.4.2.post1
